+++ src/utils/validateEmail/index.ts
@@ -1,0 +1,2 @@
+export { validateEmail } from './validateEmail';
+export type { MailtrapOptions } from './usingMailtrap';